Adding a Camera Calibration script
script outputs a text file with a series of vectors which can be used to alter your web cam image so that you can make precise measurements.
The basic idea is - your camera has a lens
lenses influence the image the computer gets - making straight lines become curved
this makes accurate measurements impossible
So you get a vector array of vectors (whatever that is) and alter the image the computer receives to make everything true to scale. So, if you know the size of something in the frame, you can use it as a reference to measure other things.