pretty background with facetracking on top
Processing code... Most of it is to get a pretty background. The redBar.pde file is responsible for the face tracking. Idea behind this was to make something that could take a simple coordinate input, and use it to add to itself. 
Using the face tracking makes the program slow, but if you comment out that part in bannerTracker.pde it runs much more smoothly. I haven't put it on the RPi yet, but I suspect it will seriously struggle with this code - it's a lot of pixel manipulation.