Update saveAlternate.py
Didn't really change much. Frame should be brighter now. That's the only actual change I made...
As for why it's slow - it's just because it's rpi I think. Your video on FB was about as fast as you can expect. Haar's cascade classifiers use a ton of processing power to find faces. It just takes time. You should be able to thread it - speeding it up, but it probably wouldn't be a huge difference.